+++ src/Command/ExecuteImportCommand.php
@@ -9,6 +9,7 @@
 namespace HeimrichHannot\EntityImportBundle\Command;
 
 use Contao\CoreBundle\Command\AbstractLockedCommand;
+use Contao\Message;
 use HeimrichHannot\EntityImportBundle\Importer\ImporterFactory;
 use HeimrichHannot\EntityImportBundle\Importer\ImporterInterface;
 use HeimrichHannot\UtilsBundle\Model\ModelUtil;
@@ -119,6 +120,14 @@ class ExecuteImportCommand extends AbstractLockedCommand
                 $this->io->success("Importer with config ID $configId finished successfully");
             } else {
                 $this->io->error("Importer with config ID $configId failed");
+
+                // transform backend messages to string
+                $messages = Message::generate();
+                $messages = str_replace('<br>', "\n", $messages);
+                $messages = strip_tags($messages);
+                Message::reset();
+
+                $this->io->error($messages);
             }
 
             if ($configModel->cronLanguage) {