+++ src/Resources/assets/js/contao-entity-import-bundle-be.js
@@ -6,12 +6,27 @@ class ContaoEntityImportBundleBe {
     }
 
     static removeWidthLimitForQuickImporters() {
-        if (document.getElementById('tl_entity_import_quick_config') === null) {
+        if (document.getElementById('tl_entity_import_quick_config') === null || document.querySelector('.list-widget') === null) {
             return;
         }
 
         document.querySelector('#header .inner').style['max-width'] = 'none';
         document.getElementById('container').style['max-width'] = 'none';
+        document.getElementById('main').style['width'] = 'auto';
+
+        let i = 0;
+
+        document.querySelectorAll('.tl_formbody_edit .widget').forEach((element) => {
+            i++;
+
+            if (null === element.querySelector('.list-widget')) {
+                element.style['max-width'] = element.classList.contains('long') ? '1000px' : '500px';
+
+                if (i % 2 === 1) {
+                    element.classList.add('clr');
+                }
+            }
+        });
     }
 }